api/pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "sevenflows-api"
version = "0.1.0"
description = "7Flows orchestration API and worker runtime."
requires-python = ">=3.11,<3.13"
dependencies = [
  "alembic>=1.14.0,<2",
  "boto3>=1.35.0,<2",
  "cryptography>=43.0.0,<45",
  "celery>=5.4.0,<6",
  "fastapi>=0.115.0,<1",
  "httpx>=0.28.0,<1",
  "pydantic-settings>=2.6.0,<3",
  "psycopg[binary]>=3.2.0,<4",
  "redis>=5.2.0,<6",
  "sqlalchemy>=2.0.36,<3",
  "uvicorn[standard]>=0.32.0,<1",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.0,<9",
  "pytest-asyncio>=0.24.0,<1",
  "ruff>=0.8.0,<1",
]

[tool.uv]
default-groups = []

[tool.setuptools.packages.find]
include = ["app*"]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]
